requirements.txt
chainlit~=2.5.5
langchain-google-genai~=2.1.3
uvicorn~=0.34.2
fastapi~=0.115.12
pydantic~=2.11.3
python-dotenv~=1.1.0
langchain~=0.3.23
langchain-google-vertexai~=2.0.20